Added serialization groups.

+++ src/AppBundle/Controller/UserController.php
@@ -5,6 +5,7 @@ namespace AppBundle\Controller;
 use AppBundle\Entity\User;
 use AppBundle\Form\UserForm;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use FOS\RestBundle\Controller\Annotations as Rest;
 use Symfony\Component\HttpFoundation\Request;
@@ -24,8 +25,11 @@ class UserController extends BaseRestController
      *
      * @Rest\View(serializerGroups={"default"})
      */
-    public function getAction(User $user)
+    public function getAction(User $user = null)
     {
+        if (!$user instanceof User) {
+            throw new NotFoundHttpException('User not found');
+        }
         return $user;
     }
 
@@ -34,6 +38,7 @@ class UserController extends BaseRestController
      * @param Request $request
      * @return \FOS\RestBundle\View\View|\Symfony\Component\Form\Form|\Symfony\Component\Form\FormInterface
      *
+     * @throws \Exception
      */
     public function postAction(Request $request)
     {
@@ -55,17 +60,46 @@ class UserController extends BaseRestController
     }
 
     /**
+     * Edit existing user.
+     * @param Request $request
+     * @param User $user
+     * @return \FOS\RestBundle\View\View|\Symfony\Component\Form\Form|\Symfony\Component\Form\FormInterface
+     *
+     * @throws \Exception
+     */
+    public function putAction(Request $request, User $user)
+    {
+        $groups = [
+            'serializerGroups' => [
+                'default',
+            ],
+            'formOptions' => [
+                'validation_groups' => [
+                    'default',
+                    'Default',
+                ],
+            ],
+        ];
+
+        return $this->handleForm($request, UserForm::class, $user, $groups, true);
+    }
+
+    /**
      * Delete user by id.
+     *
+     * @Rest\View(statusCode=204)
      *
      * @param User $user
      * @return Response
      */
-    public function deleteAction(User $user)
+    public function deleteAction(User $user = null)
     {
+        if (!$user instanceof User) {
+            throw new NotFoundHttpException('User not found');
+        }
         $em = $this->getDoctrine()->getManager();
         $em->remove($user);
         $em->flush();
-
         return null;
     }
 }
